moved enrollment-checks onto enrollment and put the email in a template
