+++ .gitignore
@@ -28,8 +28,4 @@ generated-types/
 *.cert
 *.key
 
-apps/frontend/tsconfig.node.tsbuildinfo
-apps/frontend/tsconfig.tsbuildinfo
-apps/frontend/vite.config.d.ts
-apps/frontend/vite.config.js
 prod-backup.gz
